backend-service/pyproject.toml
[tool.poetry]
name = "backend-service"
version = "0.4.2"
description = ""
authors = ["kimkun07 <kimkun07@naver.com>"]
readme = "README.md"
package-mode = false                        # Poetry only used for dependency management

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.112.0" }
google-maps-routing = "^0.6.10"
grpcio = "1.64.*"                                         # 1.65.2 version caused error BIOS data file
pydantic-extra-types = "^2.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
